feat: starting deleting some old code
