+++ src/Navigator/Actions/ActionLauncher.cs
@@ -1,5 +1,4 @@
 using System.Collections.Immutable;
-using MediatR;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Navigator.Configuration;
@@ -14,17 +13,16 @@ internal class ActionLauncher : IActionLauncher
     private readonly ILogger<ActionLauncher> _logger;
     private readonly INavigatorContextAccessor _navigatorContextAccessor;
     private readonly IServiceProvider _serviceProvider;
-    private readonly ISender _sender;
     private readonly NavigatorOptions _navigatorOptions;
     private readonly ImmutableDictionary<string, Type[]> _actions;
     private readonly ImmutableDictionary<string, ushort> _priorities;
 
-    public ActionLauncher(ILogger<ActionLauncher> logger, NavigatorOptions navigatorOptions, INavigatorContextAccessor navigatorContextAccessor, IServiceProvider serviceProvider, ISender sender)
+    public ActionLauncher(ILogger<ActionLauncher> logger, NavigatorOptions navigatorOptions,
+        INavigatorContextAccessor navigatorContextAccessor, IServiceProvider serviceProvider)
     {
         _logger = logger;
         _navigatorContextAccessor = navigatorContextAccessor;
         _serviceProvider = serviceProvider;
-        _sender = sender;
         _navigatorOptions = navigatorOptions;
         _actions = _navigatorOptions.RetrieveActions();
         _priorities = _navigatorOptions.RetrievePriorities();
@@ -43,8 +41,9 @@ internal class ActionLauncher : IActionLauncher
                 {
                     await _navigatorContextAccessor.NavigatorContext.Client.SendChatActionAsync(chatId, ChatAction.Typing);
                 }
-                
-                await _sender.Send(action);
+
+                // TODO: propagate cancellation token
+                await action.Handle(CancellationToken.None);
             }
             catch (Exception e)
             {
@@ -53,39 +52,39 @@ internal class ActionLauncher : IActionLauncher
         }
     }
 
-    private IEnumerable<IAction> GetActions()
+    private IEnumerable<Action> GetActions()
     {
         if (string.IsNullOrWhiteSpace(_navigatorContextAccessor.NavigatorContext.ActionType))
         {
-            return Array.Empty<IAction>();
+            return Array.Empty<Action>();
         }
 
         var actions = _actions
-            .Where(a => 
-                a.Key == _navigatorContextAccessor.NavigatorContext.ActionType ||
-                a.Key == nameof(ProviderAgnosticAction))
-            .ToImmutableList();
-            
+        //     .Where(a =>
+        //         a.Key == _navigatorContextAccessor.NavigatorContext.ActionType ||
+        //         a.Key == nameof(ProviderAgnosticAction))
+        .ToImmutableList();
+
         if (_navigatorOptions.MultipleActionsUsageIsEnabled())
         {
             return actions
                 .SelectMany(groups => groups.Value)
                 .SelectMany(actionType => _serviceProvider.GetServices(actionType)
-                    .Select(action =>((IAction) action!, actionType.FullName)))
+                    .Select(action => ((Action)action!, actionType.FullName)))
                 .Where(a => a.Item1.CanHandleCurrentContext())
                 .OrderBy(a => _priorities.GetValueOrDefault(a.FullName ?? string.Empty, Priority.Default))
                 .Select(a => a.Item1)
                 .AsEnumerable();
         }
-            
+
         var action = actions
             .SelectMany(groups => groups.Value)
             .SelectMany(actionType => _serviceProvider.GetServices(actionType)
-                .Select(action =>((IAction) action!, actionType.FullName)))
+                .Select(action => ((Action)action!, actionType.FullName)))
             .OrderBy(a => _priorities.GetValueOrDefault(a.FullName ?? string.Empty, Priority.Default))
             .Select(a => a.Item1)
             .FirstOrDefault(a => a.CanHandleCurrentContext());
 
-        return action is not null ? new[] {action} : Array.Empty<IAction>();
+        return action is not null ? new[] { action } : [];
     }
 }
